fix: travis push when tag exists

+++ .travis/push.sh
@@ -13,18 +13,21 @@ echo "生成压缩文件"
 export GZ_FILE=wechat-devtools-$version.tar.xz
 tar -cJf $GZ_FILE dist --transform 's/^dist/wechat-devtools-'$version'/'
 
+git remote set-url origin https://sigoden:${GT_TOKEN}@github.com/sigoden/wechat-devtools.git
+
 echo "提交 git commit"
 if $(git diff --name-only | grep -qx version); then # 有新版本
     git checkout master
     git add version
     git commit -m "ci: upgrade to $version; [skip ci]"
     git tag $version
-    git remote set-url origin https://sigoden:${GT_TOKEN}@github.com/sigoden/wechat-devtools.git
     git push origin master --tags
 else
-    git tag -f $version
-    git remote set-url origin https://sigoden:${GT_TOKEN}@github.com/sigoden/wechat-devtools.git
-    git push --delete origin $version
+    git fetch --tags
+    git tag $version > /dev/null 2>&1
+    if [ "$?" -ne "0" ]; then
+        git push --delete origin $version
+    fi
     git push origin $version
 fi
